cpp-08 cpp-09 bug fixed

+++ cpp-08/ex00/main.cpp
@@ -14,13 +14,13 @@ int main() {
 	v.push_back(3);
 	v.push_back(2);
 	v.push_back(1);
-	   for (int x = 0; x < v.size(); x++) {
+	   for (size_t x = 0; x < v.size(); x++) {
         std::cout << &*i << " - " << *i++ << " " << std::endl;
     }
 	v.push_front(6);
 	std::cout << "------" << std::endl;
 	i = v.begin();
-	for (int x = 0; x < v.size(); x++) {
+	for (size_t x = 0; x < v.size(); x++) {
         std::cout << &*i << " - " << *i++ << " " << std::endl;
     }
 	return 0;
